requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.39.0
python-multipart==0.0.9
requests==2.32.5

psycopg2-binary==2.9.10
SQLAlchemy==2.0.36

qdrant-client==1.13.2
sentence-transformers==3.2.1
transformers==4.46.3
torch==2.9.1
sentencepiece==0.2.1

pytesseract==0.3.13
python-docx==1.1.2
PyPDF2==3.0.1
pdfplumber==0.11.0
Pillow
